+++ src/pages/dex/Dex.tsx
@@ -1,11 +1,11 @@
-import PokemonList from '@/components/dex/PokemonList';
+import PokemonDex from '@/components/dex/PokemonDex';
 import FilterPlates from '@/components/plate/FilterPlates';
 
 const Dex = () => {
   return (
     <section>
       <FilterPlates />
-      <PokemonList />
+      <PokemonDex />
     </section>
   );
 };
